requirements.txt
accelerate==0.27.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
diffusers==0.26.3
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.0.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tifffile==2024.8.30
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
